fix(starlight-karma-bot): bind health server before Discord login
